+++ menus-back/src/main/java/fr/choupiteam/menus/application/side/model/SideDish.java
@@ -30,7 +30,7 @@ public class SideDish {
      */
     @JsonSerialize(using = IngredientMapSerializer.class)
     @JsonDeserialize(using = IngredientMapDeserializer.class)
-    private Map<String, Integer> ingredients;
+    private Map<String, Float> ingredients;
 
     public String getId() {
         return id;
@@ -48,11 +48,11 @@ public class SideDish {
         this.name = name;
     }
 
-    public Map<String, Integer> getIngredients() {
+    public Map<String, Float> getIngredients() {
         return ingredients;
     }
 
-    public void setIngredients(Map<String, Integer> ingredients) {
+    public void setIngredients(Map<String, Float> ingredients) {
         this.ingredients = ingredients;
     }
 }
